+++ package.json
@@ -15,7 +15,6 @@
     "@graphql-codegen/cli": "1.2.1-alpha-3d654462.34",
     "@graphql-codegen/typescript": "1.2.1-alpha-3d654462.34",
     "@graphql-codegen/typescript-resolvers": "1.2.1-alpha-3d654462.34",
-    "@types/axios": "^0.14.0",
     "@types/bcrypt": "3.0.0",
     "@types/body-parser": "1.17.0",
     "@types/cookie": "0.3.3",
@@ -35,9 +34,9 @@
   },
   "dependencies": {
     "@safe-api/middleware": "0.0.2",
+    "apollo-datasource-rest": "0.5.0",
     "apollo-server-express": "2.6.0",
     "apollo-server-testing": "2.6.0",
-    "axios": "0.19.0",
     "bcrypt": "3.0.6",
     "body-parser": "1.19.0",
     "cookie": "0.4.0",
